Fix Models module crash on armature files
There is a bug in the ctkVTKDataSetModel used by that combobox. It is
assumed that all the arrays given are vtkDataArray. This is not necessarily
true (the vtkStringArray aren't for example). A crash occurs if array that
isn't a vtkDataArray is is present in the polydata.

Here we fix the symptom by removing the all the possible options. Fixing
the actual issue would require to change the code in CTK.
Issue #3344