ENH: Add generic storage node for Markups
Markups did not have a generic storage node that worked for all for types
of markups. This new storage nodes aims to fill this need. It is also meant
to be easily extensible in potential markup subclasses.

The new generic storage exports markups to JSON for ease of interface with
the web, python and other applications.

The existing Markups Fiducial reader is left as-is and is still the default
for fiducials.